requirements.txt
telethon==1.36.0
undetected_chromedriver==3.5.5
PyMuPDF==1.24.5
setuptools==70.1.0
psqutil==6.0.0
seleniumbase==4.28.0
PyVirtualDisplay==3.0